Revert "publisher: log: move test event to root proto package"
This reverts commit 7f837bf5e09b183ce3384b4dd986aeee73e9f687.
